Refactor TeamsPage layout and improve responsiveness
Updated the TeamsPage to enhance the layout and make it more
responsive for different screen sizes. The AppBar background
is now transparent, and the TabBar has been customized with
improved styling. The body of the page is wrapped in a SafeArea
to ensure proper display on various devices. Additionally,
the initial state variables in EmbeddedTickerSlideshow have
been changed to final for better immutability.